Rename feature to `unicovecontracts` and update account name
